Add abstraction for provider thing IDs
It is not always possible to have the same IDs on each provider.
Leaderboard IDs and achievement IDs are now abstracted from the consumer
of the GameKit plugin. The IDs returned from getLeaderboards and passed
into submitScore are now a reference id which gets mapped to the correct
provider ID. It is the responsibility of the developer to properly
configure the mapping.